+++ src/Bladeburner.js
@@ -3185,7 +3185,7 @@ function initBladeburner() {
     Skills[SkillNames.Overclock] = new Skill({
         name:SkillNames.Overclock,
         desc:"Each level of this skill decreases the time it takes " +
-             "to attempt a contract or operation by 1% (Max Level: 99)",
+             "to attempt a contract or operation by 1% (Max Level: 95)",
         baseCost:5, costInc:1, maxLvl:95,
         actionTime:1
     });