+++ src/Player.js
@@ -19,7 +19,7 @@ import {AllServers, Server, AddToAllServers}    from "./Server.js";
 import {SpecialServerIps, SpecialServerNames}   from "./SpecialServerIps.js";
 import {SourceFiles, applySourceFile}           from "./SourceFile.js";
 
-import Decimal                                  from '../utils/decimal.js';
+import Decimal                                  from "decimal.js";
 import {dialogBoxCreate}                        from "../utils/DialogBox.js";
 import {clearEventListeners}                    from "../utils/HelperFunctions.js";
 import {createRandomIp}                         from "../utils/IPAddress.js";
@@ -99,6 +99,7 @@ function PlayerObject() {
     this.currentServer          = ""; //IP address of Server currently being accessed through terminal
     this.purchasedServers       = []; //IP Addresses of purchased servers
     this.hacknetNodes           = [];
+    this.hacknetNodeWrappers    = [];
     this.totalHacknetNodeProduction = 0;
 
     //Factions
@@ -390,15 +391,7 @@ PlayerObject.prototype.prestigeSourceFile = function() {
     if (this.bitNodeN === 3) {this.money = new Decimal(150e9);}
     this.corporation = 0;
 
-    //Reset Bladeburner
     this.bladeburner = 0;
-
-    //BitNode 8: Ghost of Wall Street
-    if (this.bitNodeN === 8) {this.money = new Decimal(100000000);}
-    if (this.bitNodeN === 8 || hasWallStreetSF) {
-        this.hasWseAccount = true;
-        this.hasTixApiAccess = true;
-    }
 
     this.playtimeSinceLastAug = 0;
     this.scriptProdSinceLastAug = 0;